// website/sanity.types.ts
export type QueryKitsResult = Array<{
  _createdAt: string;
  _id: string;
  name: string | null;
  preset: string | null;
  description: string | null;
  repo: string | null;
  status: "approved" | "pending" | "rejected" | null;
  votes: number | null;
  owner: {
    name: string | null;
    username: string | null;
    githubId: string | null;
    avatarUrl: string | null;
  } | null;
}>;